# gateway-service/json/api-doc.json
{"swagger":"2.0","info":{"description":"Documentation Financial Institute API v1.0","version":"1.0","title":"Financial API"},"host":"localhost:8092","basePath":"/","tags":[{"name":"account-controller","description":"Account Controller"},{"name":"financial-institute-controller","description":"Financial Institute Controller"},{"name":"meta-data-contoller","description":"Meta Data Contoller"},{"name":"transaction-controller","description":"Transaction Controller"}],"paths":{"/account/{cust_id}":{"get":{"tags":["account-controller"],"summary":"get","operationId":"getUsingGET","produces":["*/*"],"parameters":[{"name":"cust_id","in":"path","description":"cust_id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Account"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["account-controller"],"summary":"add","operationId":"addUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"account","description":"account","required":true,"schema":{"$ref":"#/definitions/Account"}},{"name":"cust_id","in":"path","description":"cust_id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Account"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/fi":{"get":{"tags":["financial-institute-controller"],"summary":"get","operationId":"getUsingGET_1","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/FinancialInstitute"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["financial-institute-controller"],"summary":"add","operationId":"addUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"fi","description":"fi","required":true,"schema":{"$ref":"#/definitions/FinancialInstitute"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/FinancialInstitute"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/fi/{fiID}/endpoints":{"get":{"tags":["financial-institute-controller"],"summary":"getEndPoints","operationId":"getEndPointsUsingGET","produces":["*/*"],"parameters":[{"name":"fiID","in":"path","description":"fiID","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/EndPoints"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/fi/{fiID}/endpoints/{operation}":{"post":{"tags":["financial-institute-controller"],"summary":"addEndPoint","operationId":"addEndPointUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"endPoint","description":"endPoint","required":true,"schema":{"$ref":"#/definitions/EndPoint"}},{"name":"fiID","in":"path","description":"fiID","required":true,"type":"integer","format":"int64"},{"name":"operation","in":"path","description":"operation","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/EndPoint"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/fi/{fiID}/metadata":{"get":{"tags":["meta-data-contoller"],"summary":"get","operationId":"getUsingGET_2","produces":["*/*"],"parameters":[{"name":"fiID","in":"path","description":"fiID","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/MetaDef"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["meta-data-contoller"],"summary":"add","operationId":"addUsingPOST_2","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"fiID","in":"path","description":"fiID","required":true,"type":"integer","format":"int64"},{"in":"body","name":"meta","description":"meta","required":true,"schema":{"$ref":"#/definitions/MetaDef"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/MetaDef"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/fi/{id}":{"get":{"tags":["financial-institute-controller"],"summary":"findById","operationId":"findByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/FinancialInstitute"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/fi/{id}/endpoints":{"post":{"tags":["financial-institute-controller"],"summary":"addEndPoints","operationId":"addEndPointsUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"endPoints","description":"endPoints","required":true,"schema":{"$ref":"#/definitions/EndPoints"}},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/EndPoints"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/transaction/{cust_id}":{"get":{"tags":["transaction-controller"],"summary":"get","operationId":"getUsingGET_3","produces":["*/*"],"parameters":[{"name":"cust_id","in":"path","description":"cust_id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"array","items":{"$ref":"#/definitions/Transaction"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["transaction-controller"],"summary":"add","operationId":"addUsingPOST_3","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"cust_id","in":"path","description":"cust_id","required":true,"type":"integer","format":"int64"},{"in":"body","name":"transaction","description":"transaction","required":true,"schema":{"$ref":"#/definitions/Transaction"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Transaction"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"Account":{"type":"object","properties":{"accountKey":{"type":"string"},"accountNumber":{"type":"string"},"accountType":{"type":"string","enum":["CHECKING","CURRENT","SAVINGS"]},"bankID":{"type":"string"},"branchID":{"type":"string"},"customerID":{"type":"integer","format":"int64"}},"title":"Account"},"EndPoint":{"type":"object","properties":{"endPointType":{"type":"string","enum":["LINK","UNLINK","TRANSACTION","STATEMENT","TAX_STATEMENT"]},"serviceDef":{"$ref":"#/definitions/ServiceDef"},"type":{"type":"string","enum":["API","PROVIDED","WEB","OFX"]}},"title":"EndPoint"},"EndPoints":{"type":"object","properties":{"desc":{"type":"string"},"endpoints":{"type":"array","items":{"$ref":"#/definitions/EndPoint"}},"fiID":{"type":"integer","format":"int64"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"}},"title":"EndPoints"},"FinancialInstitute":{"type":"object","properties":{"address":{"type":"string"},"authType":{"type":"string","enum":["PASSWORD","OAUTH"]},"id":{"type":"integer","format":"int64"},"intMethod":{"type":"string","enum":["API","PROVIDED","WEB","OFX"]},"name":{"type":"string"}},"title":"FinancialInstitute"},"MetaDef":{"type":"object","properties":{"fiID":{"type":"integer","format":"int64"},"fields":{"type":"object","additionalProperties":{"type":"string"}},"objectType":{"type":"string","enum":["Account","Transaction"]}},"title":"MetaDef"},"ServiceDef":{"type":"object","properties":{"acceptContent":{"type":"string"},"contentType":{"type":"string"},"errors":{"type":"array","items":{"type":"integer","format":"int32"}},"method":{"type":"string"},"url":{"type":"string"}},"title":"ServiceDef"},"Transaction":{"type":"object","properties":{"accountID":{"type":"integer","format":"int64"},"amount":{"type":"number"},"bankID":{"type":"integer","format":"int64"},"id":{"type":"string"},"transactionDate":{"type":"string","format":"date-time"},"type":{"type":"string","enum":["CREDIT","DEBIT","INT","DIV","FEE","SRVCHG"]}},"title":"Transaction"}}}